generated_projects/multi_city_weather_comparison/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
mypy==1.5.1
tenacity==8.2.3
prometheus-client==0.17.1